requirements.txt
PySide6>=6.6.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
